recipes_project/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.20
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.2
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.20.1
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
